+++ src/node/utils/ExportEtherpad.js
@@ -55,5 +55,10 @@ exports.getPadRaw = async (padId, readOnlyId) => {
   })();
   const data = {[dstPfx]: pad};
   for (const [dstKey, p] of new Stream(records).batch(100).buffer(99)) data[dstKey] = await p;
+  await hooks.aCallAll('exportEtherpad', {
+    pad,
+    data,
+    dstPadId: readOnlyId || padId,
+  });
   return data;
 };